backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
